+++ src/export_1.c
@@ -12,6 +12,31 @@
 
 #include "minishell.h"
 
+void	change_environment_variables_question_mark(int value, t_data *data)
+{
+	int		i;
+	int		len;
+	char	*new_env;
+	char	*value_str;
+
+	i = 0;
+	while (data->envp[i])
+		i++;
+	if (ft_atoi(data->envp[i - 1] + 2) != value)
+	{
+		len = ft_nblen(value, 10);
+		new_env = ft_calloc(3 + len, sizeof(char));
+		if (!new_env)
+			return ;
+		ft_strlcpy(new_env, "?=", 3);
+		value_str = ft_itoa(value);
+		new_env = ft_strjoin_free(new_env, value_str);
+		free(data->envp[i - 1]);
+		data->envp[i - 1] = new_env;
+		free(value_str);
+	}
+}
+
 void	export(t_data *data)
 {
 	int			i1;
