+++ src/export_3.c
@@ -12,114 +12,58 @@
 
 #include "minishell.h"
 
-static void	print_env_order(t_data *data)
+static char	*reorganization(char *env, int len)
 {
-	int		i1;
-	int		i2;
-	char	alpha;
-
-	i1 = 0;
-	alpha = 'a';
-	while (i1 < 26)
-	{
-		i2 = 0;
-		while (data->envp[i2])
-		{
-			if (data->envp[i2][0] == alpha || data->envp[i2][0] == alpha - 32)
-				ft_printf("declare -x %s\n", data->envp[i2]);
-			i2++;
-		}
-		alpha++;
-		i1++;
-	}
-	i2 = 0;
-	while (data->envp[i2])
-	{
-		if (ft_isalpha(data->envp[i2][0]) == 0 &&
-				!ft_strchr(data->envp[i2], '?'))
-			ft_printf("declare -x %s\n", data->envp[i2]);
-		i2++;
-	}
-}
-
-bool	print_export(t_data *data)
-{
-	int			i1;
-	t_new_list	*aux;
-
-	i1 = 0;
-	aux = data->list;
-	while (aux->content[i1])
-		i1++;
-	if (!ft_strncmp(aux->content[i1 - 1], "export",
-			ft_strlen(aux->content[i1 - 1])))
-	{
-		print_env_order(data);
-		change_environment_variables_question_mark(0, data);
-		return (true);
-	}
-	return (false);
-}
-
-static void	check_character(int i1, bool *add_var, t_new_list *aux)
-{
-	int	i2;
-
-	i2 = 0;
-	while (aux->content[i1][i2])
-	{
-		if (aux->content[i1][i2] == '=' && ft_strlen(aux->content[i1]) == 1)
-		{
-			write(2, "export: `=': not a valid identifier\n", 36);
-			*add_var = false;
-			write(2, "\n", 1);
-			break ;
-		}
-		if (!ft_isalnum(aux->content[i1][i2])
-			&& aux->content[i1][i2] != '_' && aux->content[i1][i2] != '='
-			&& aux->content[i1][i2] != '\'' && aux->content[i1][i2] != '\"'
-			&& aux->content[i1][i2] != '$' && aux->content[i1][i2] != '\\'
-			&& aux->content[i1][i2] != ' ' && aux->content[i1][i2] != '.'
-			&& aux->content[i1][i2] != '/' && aux->content[i1][i2] != ':'
-			&& aux->content[i1][i2] != '-' && aux->content[i1][i2] != '+')
-		{
-			write(2, "export: not an identifier: ", 27);
-			ft_putstr_fd(aux->content[i1], 2);
-			write(2, "\n", 1);
-			*add_var = false;
-			break ;
-		}
-		i2++;
-	}
-}
-
-bool	check_error(int i1, t_new_list *aux)
-{
-	bool	add_var;
-
-	add_var = true;
-	if (!ft_isalpha(aux->content[i1][0]) && aux->content[i1][0] != '_')
-	{
-		write(2, "export: `", 9);
-		ft_putstr_fd(aux->content[i1], 2);
-		write(2, "': not a valid identifier\n", 26);
-		add_var = false;
-	}
-	else
-		check_character(i1, &add_var, aux);
-	return (add_var);
-}
-
-int	ft_strnchrcmp(const char *s1, const char *s2, size_t n, char chr)
-{
-	size_t	i;
+	int		i;
+	char	*new_env;
 
 	i = 0;
-	while ((s1[i] || s2[i]) && (i < n) && s1[i] != chr)
+	new_env = (char *) malloc(len + 3);
+	if (!new_env)
+		return (NULL);
+	i = 0;
+	while (env[i] != '=')
 	{
-		if (s1[i] != s2[i])
-			return (s1[i] - s2[i]);
+		new_env[i] = env[i];
 		i++;
 	}
+	new_env[i] = env[i];
+	i++;
+	new_env[i] = '"';
+	i++;
+	ft_strlcpy(new_env + i, env + i - 1, len);
+	new_env[len + 1] = '"';
+	new_env[len + 2] = '\0';
+	return (new_env);
+}
+
+char	*analize_env(char *env)
+{
+	int	i;
+	int	len;
+	int	count_equals;
+
+	i = 0;
+	len = ft_strlen(env);
+	count_equals = 0;
+	while (i < len)
+	{
+		if (env[i] == '=')
+			count_equals++;
+		i++;
+	}
+	if (count_equals > 1 || env[len - 1] == '=')
+	{
+		return (reorganization(env, len));
+	}
+	return (ft_strdup(env));
+}
+
+int	is_valid_char(char c)
+{
+	if (c == '_' || c == '=' || c == '\'' || c == '\"' || c == '$' || c == '\\')
+		return (1);
+	if (c == ' ' || c == '.' || c == '/' || c == ':' || c == '-' || c == '+')
+		return (1);
 	return (0);
 }
