+++ src/main.c
@@ -71,7 +71,7 @@ void	master(char *command, t_data *data)
 		else if (!ft_strncmp(aux->content[i], "cd", ft_strlen(aux->content[i])))
 			cd(aux, data);
 		else if (!ft_strncmp(aux->content[i], "echo", ft_strlen(aux->content[i])))
-			echo(data);
+			echo(aux, data);
 		else if (!ft_strncmp(aux->content[i], "env", ft_strlen(aux->content[i])))
 			env(aux, data);
 		else if (!ft_strncmp(aux->content[i], "export", ft_strlen(aux->content[i])))
